+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## {PR_MERGE_DATE}
 
-- Initial release with Latest Teardowns and Teardown of the Day commands.
+- Initial release with Latest Teardowns, Teardown of the Day, and Random Business Idea commands.